Set GSM Network auth-policy to closed by default.
We don't want to ship a live system that would bring up an accept-all network!

This commit also removes the auth-policy option from the install script.
If the user cannot manually intervene to modify this, then they certainly
should not be running an accept-all network.

Also, removed seemingly pointless grep commands and changed
the default password for sed to match on. (pending fixing all that correctly)